ai-python-service/requirements.txt
fastapi
uvicorn
google-generativeai==0.3.0
celery
redis
pydantic
python-dotenv
requests
